# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Netflix Top 10 Stremio Addon API",
    "description": "API for Netflix Top 10 rankings per-country with Stremio integration. Provides live Netflix Top 10 rankings scraped from FlixPatrol with TMDB metadata matching.",
    "version": "3.7.0",
    "contact": {
      "name": "Netflix Top 10 Addon",
      "url": "https://github.com/C0st1/top10addon"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/api/validate-tmdb-key": {
      "post": {
        "tags": ["API"],
        "summary": "Validate TMDB API key",
        "description": "Validates a TMDB API key by making a test request to the TMDB API",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["apiKey"],
                "properties": {
                  "apiKey": {
                    "type": "string",
                    "description": "TMDB API key (v3 or v4)",
                    "example": "8a7f3bc2d1e4f5a6b7c8d9e0f1a2b3c4"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "X-RateLimit-Remaining": {
                "schema": { "type": "integer" },
                "description": "Remaining requests in current window"
              },
              "X-RateLimit-Reset": {
                "schema": { "type": "integer" },
                "description": "Seconds until window resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable (circuit breaker open)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/save-config": {
      "post": {
        "tags": ["API"],
        "summary": "Save configuration",
        "description": "Saves user configuration and returns an opaque token for the manifest URL. The token does not contain sensitive data.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigRequest": {
        "type": "object",
        "required": ["tmdbApiKey"],
        "properties": {
          "tmdbApiKey": {
            "type": "string",
            "description": "TMDB API key",
            "minLength": 20,
            "maxLength": 200
          },
          "rpdbApiKey": {
            "type": "string",
            "description": "RPDB API key (optional)"
          },
          "country": {
            "type": "string",
            "description": "Comma-separated country names",
            "default": "Global",
            "example": "Global,United States,Japan"
          },
          "movieType": {
            "type": "string",
            "description": "Custom type for movies tab",
            "default": "movie"
          },
          "seriesType": {
            "type": "string",
            "description": "Custom type for series tab",
            "default": "series"
          }
        }
      },
      "ConfigResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque token for manifest URL",
            "example": "A1B2C3D4E5F6G7H8I9J0K1L2M3N4O5P6"
          },
          "manifestUrl": {
            "type": "string",
            "description": "Full URL to manifest.json",
            "example": "https://example.com/A1B2C3D4E5F6G7H8I9J0K1L2M3N4O5P6/manifest.json"
          },
          "installUrl": {
            "type": "string",
            "description": "Stremio install URL",
            "example": "stremio://example.com/A1B2C3D4E5F6G7H8I9J0K1L2M3N4O5P6/manifest.json"
          }
        }
      },
      "ValidationResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "description": "Request ID for debugging (only on server errors)"
          }
        }
      },
...
    }
  }
}